Fixed bug that caused euclidean distance to be calculated incorrectly when array type was np.uint8

+++ utils/mnist_reader.py
@@ -14,4 +14,4 @@ def load_mnist(path, kind='train'):
     with gzip.open(images_path, 'rb') as imgpath:
         images = np.frombuffer(imgpath.read(), dtype=np.uint8, offset=16).reshape(len(labels), 784)
 
-    return images, labels
+    return images.astype('float32'), labels.astype('float32')
